fix(rfc64): centralize selected activation
